% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{nll}
\alias{nll}
\title{Calculate Negative Log-Likelihood for AFT Models}
\usage{
nll(Y, eta, delta, sigma, n, model)
}
\arguments{
\item{Y}{Vector of observed times (response variable)}

\item{eta}{Vector of linear predictors}

\item{delta}{Vector of censoring indicators (1 for event, 0 for censored)}

\item{sigma}{Scale parameter}

\item{n}{Sample size}

\item{model}{Character string specifying the distribution ("weibull", "lognormal", or "loglogistic")}
}
\value{
Normalized negative log-likelihood value (i.e., divided by sample size)
}
\description{
Computes the negative log-likelihood for Accelerated Failure Time (AFT) models,
supporting Weibull, lognormal, and loglogistic distributions.
}
\note{
Last change 15/03/2025
}
\seealso{
\code{\link{ll}} for the positive log-likelihood version
}
\keyword{internal}
