% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aft_coop.cv.R
\name{plot_aft_coop.cv}
\alias{plot_aft_coop.cv}
\title{Plot AFT Cooperative Cross-Validation Results}
\usage{
plot_aft_coop.cv(cv.out, nfolds, case, addbar = F)
}
\arguments{
\item{cv.out}{List containing cross-validation results obtained by \code{\link{cv.coop}} , must include components:
\itemize{
\item cv.err.obj - Matrix of CV errors for each fold and lambda value
\item cv.err.linPred - Vector of mean CV errors for linear predictor
\item  lambda_grid Numeric vector of lambda values used in the regularization path.
\item lambda.1se - Lambda value within 1 standard error of minimum CV error
\item lambda.min - Lambda value with minimum CV error
\item case.rho - Rho value for the current case
}}

\item{nfolds}{Integer specifying the number of cross-validation folds used.}

\item{case}{Character or numeric identifier for the current case being analyzed.}

\item{addbar}{Logical to add error bar to the crossvalidation curve (Default FALSE)}
}
\value{
A ggplot object visualizing cross-validation results
}
\description{
Internal function that creates a ggplot visualization of cross-validation results
from an accelerated failure time (AFT) cooperative regularization model.
}
\details{
The function calculates standard errors for cross-validation error estimates
and creates a visualization with error bars. It marks the lambda.min and lambda.1se
values with vertical dashed lines. The object cv.err.obj is obtained using \code{\link{cv.coop}}
}
\note{
Last change 15/04/2025
}
\keyword{internal}
