% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{prox.l1}
\alias{prox.l1}
\title{L1 Proximal Operator}
\usage{
prox.l1(u, mu)
}
\arguments{
\item{u}{Numeric vector of input values}

\item{mu}{Threshold parameter (typically lambda * alpha / L)}
}
\value{
Vector of proximal operator values
}
\description{
Computes the proximal operator for L1 regularization (soft thresholding).
}
\details{
Implements the soft thresholding operator:
\deqn{prox_{\mu\|\cdot\|_1}(u) = sign(u)\max(|u| - \mu, 0)}
}
\note{
Last change 11/03/2025
}
\keyword{internal}
