% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{standardize}
\alias{standardize}
\title{Standardize Matrix Columns}
\usage{
standardize(mat)
}
\arguments{
\item{mat}{Input matrix to be standardized}
}
\value{
A list containing:
\itemize{
\item std: Standardized matrix
\item means: Vector of column means
\item sds: Vector of column standard deviations
}
}
\description{
Standardizes a matrix by centering and scaling each column to have zero mean
and unit variance.
}
\note{
Last change 11/03/2025
}
\examples{
\dontrun{
X <- matrix(rnorm(100), 20, 5)
std_result <- standardize(X)
# Access standardized matrix
X_std <- std_result$std
# Original scale restoration
X_restored <- scale(X_std, center = -std_result$means, scale = 1/std_result$sds)
}

}
\keyword{internal}
